requirements.txt
Django
djangorestframework
django-filter
requests
